requirements.txt
streamlit
spotipy
pandas
scikit-learn
python-dotenv
